Migrate backlog into a component
